Half-fix slider groove being drawn twice
When drawComplexControl draws SC_SliderGroove, it draws the groove and
then highlights an area from minimum() to the current value.

Previously this code was setting the position/value to 0 to avoid
drawing the highlight, but since we're setting minimum() to below zero
it was drawing from -100 to 0 and looking odd.

This fix sets the position to minimum() to avoid that.

There is one leftover issue where a small dot is drawn on the far left
under the Gtk+ style, but I can't see an easy way to fix that.