rotator: Set the correct phase at the start of a block
Previously the phase of the rotator depended on state of the previous run.

This could've been from any block so it was unlikely to be continuous and
led to the phase of the filtered signal changing on every repaint and
introduced discontinuities between block boundaries.

This change passes the first sample ID into the flowgraph as a tag and
the rotator uses this to set the initial phase.