Display drag/move mouse pointer when dragging cursors
Enables mouse tracking on PlotView to get mouse move events
even when not dragging.

Passes through Leave events to handle the case where the cursor
is near the edge of the widget and the mouse leaves the widget
without generating a mouse move event that isn't on the cursor.

Passes in mouse cursor shape to Cursor to define whether it should
be a horizontal resize, vertical resize or move (resize all).

Qt handles the case where the cursor is dragged off screen and
does not generate a Leave event while draggin (<3 Qt).