Force fft/zoom change signal to setup initial state
This wasn't a problem before as setDefaults would force
a changed signal. Now with settings saving it's possible
for the saved setting to be the same as the initial slider
position and a changed event is not emitted.

This leads to a case where the rest of the app doesn't get
synced up and causes a crash.

Fixes #109