Display annotation comments as tooltips
When the mouse is moved over a SigMF annotation rectangle on the
waterfall, a tooltip showing the annotation comment is displayed
(if the annotation has a comment field).

This feature can be enabled/disabled with a checkbox, in the same
way as annotations. If annotations are disabled, the annotation comments
checkbox is grayed out and the tooltips are not shown.

Signed-off-by: Daniel Estévez <daniel@destevez.net>