Include format specifiers in help text
Previously, the format specifiers were only included in the source
repository README and the source code; this change adds the specifiers
to the program help text as well.

I looked at the QCommandLineOption API and there does not seem to be a
way to specify a subset of allowed strings such that they are
automatically included in the help text by QCommandLineParser - the
canonical way is just to include the options in the matching help text
as I have done in this commit.