sigmf: Add support for datasets
Sometimes, we want to load a data file found under the `core:dataset`
key. In these cases, we need to read the dataset key and overwrite the
respective variable.
Further, we'd need to check if the corresponding `sigmf-data` file
exists, because it is supposed to take precedence over the dataset file.
The dataset file is all called a Non-Conforming-Datafile (NCD).

Signed-off-by: Johannes Demel <johannes.demel@dedrone.com>